+++ Dockerfile_test_build
@@ -1,0 +1,30 @@
+# the actual image
+FROM ubuntu:16.04
+
+WORKDIR /srv
+
+# install all the system dependancies
+RUN apt-get update && apt-get install -y python2.7 \
+        python-pip fping python-dev \
+        libpq-dev libffi-dev unixodbc unixodbc-dev \
+        freetds-bin freetds-dev tdsodbc strace htop \
+        inetutils-ping nmap traceroute netcat tcpdump \
+        net-tools bmon entr git pkg-config nano vim wget \
+        curl supervisor virtualenv autoconf libsnmp-dev \
+        libncurses5-dev g++ cmake libssl-dev libaio1 liblz4-1
+
+# install Go1.9
+RUN mkdir /srv/staging
+RUN wget -O /srv/staging/go1.13.linux-amd64.tar.gz https://dl.google.com/go/go1.13.linux-amd64.tar.gz
+RUN tar zxvf /srv/staging/go1.13.linux-amd64.tar.gz -C /srv
+RUN ln -s /srv/go/bin/go /usr/local/bin/go
+RUN rm /srv/staging/go1.13.linux-amd64.tar.gz
+
+# create and update both virtual environments; but don't install anything yet
+RUN mkdir /srv/Envs \
+    && virtualenv /srv/Envs/backend_py \
+    && . /srv/Envs/backend_py/bin/activate \
+    && pip install --upgrade pip setuptools \
+    && deactivate
+
+ENV TERM="xterm" GODEBUG="cgocheck=0"
