+++ README.md
@@ -113,3 +113,9 @@ SNMPVariable(
 Some of this may feel a bit like [easysnmp](https://github.com/kamakazikamikaze/easysnmp); that's intentional, I was originally using that but I think its got some underlying thread-safety issues on the C side (particularly to do with SNMPv3).
 
 No offence to the guys that contribute to that project- it's served me very well.
+
+
+To use the test build container...
+
+    docker build --tag gosnmp_python_test_build -f Dockerfile_test_build .
+    docker run --rm -it -v `pwd`:/workspace gosnmp_python_test_build bash
