+++ go.mod
@@ -1,0 +1,11 @@
+module gosnmp_python
+
+go 1.13
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/ftpsolutions/gosnmp v0.0.0-20180913035423-cdbbd56a591c
+	github.com/go-python/gopy v0.3.1
+	github.com/stretchr/testify v1.4.0 // indirect
+	golang.org/x/tools v0.0.0-20190910044552-dd2b5c81c578 // indirect
+)
