+++ gosnmp_python_go/utils.go
@@ -1,4 +1,4 @@
-package gosnmp_python
+package gosnmp_python_go
 
 // #cgo pkg-config: python2
 // #include <Python.h>
